refactor of modules, plus tests
